requirements.txt
Pillow~=5.1.0
werkzeug~=0.14.0